requirements.txt
# Essential Flask Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests for APIs
requests==2.31.0

# Image Processing
Pillow==10.1.0
numpy==1.25.2

# Data Processing
pandas==2.1.4
scikit-learn==1.3.2

# Production Web Server
gunicorn==21.2.0

# OpenWeatherMap Client (Alternative)
pyowm==3.3.0

# Additional Flask utilities
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
